Player not moving fix.
git-svn-id: file:///home/cbm/dev/rpgtoolkit/svn/trunk@1653 8ad37bdc-5837-4682-a52a-fbb6d9f1e9f7

+++ vc/trans3/movement/CSprite/CSprite.cpp
@@ -242,7 +242,7 @@ bool CSprite::push(const bool bScroll)
 	extern LPBOARD g_pBoard;
 
 	// Pixels per frame (tile or pixel movement - not moveSize()).
-	const double stepSize = PX_FACTOR / m_pos.loopSpeed;
+	const double stepSize = double(PX_FACTOR) / m_pos.loopSpeed;
 	// Integer values to scroll.
 	int scx = -round(m_pos.x), scy = -round(m_pos.y);
